+++ src/Rearch/ListenerHandle.cs
@@ -6,13 +6,13 @@ namespace Rearch;
 
 /// <summary>
 /// A handle onto the lifecycle of a listener from
-/// <see cref="Container.Listen(CapsuleListener)"/>.
+/// <see cref="CapsuleContainer.Listen(CapsuleListener)"/>.
 /// You <i>must</i> <see cref="Dispose()"/> the <see cref="ListenerHandle"/>
 /// when you no longer need the listener in order to prevent leaks.
 /// </summary>
 public class ListenerHandle : IDisposable
 {
-    private readonly Container container;
+    private readonly CapsuleContainer container;
     private readonly Capsule<object?> capsule;
 
     /// <summary>
@@ -20,7 +20,7 @@ public class ListenerHandle : IDisposable
     /// </summary>
     /// <param name="container">Container containing listener's data.</param>
     /// <param name="capsule">Capsule whose data to listen to.</param>
-    internal ListenerHandle(Container container, Capsule<object?> capsule)
+    internal ListenerHandle(CapsuleContainer container, Capsule<object?> capsule)
     {
         this.container = container;
         this.capsule = capsule;
@@ -43,7 +43,11 @@ public class ListenerHandle : IDisposable
     {
         if (disposing)
         {
-            this.container.Manager(this.capsule).Dispose();
+            var capsules = this.container.Capsules;
+            if (capsules.ContainsKey(this.capsule))
+            {
+                capsules[this.capsule].Dispose();
+            }
         }
     }
 }