+++ src/Rearch/UntypedCapsuleManager.cs
@@ -7,45 +7,41 @@ namespace Rearch;
 /// <summary>
 /// Base untyped dataflow node that manages capsule data and effects.
 /// </summary>
-internal abstract class UntypedCapsuleManager(Container container) :
+internal abstract class UntypedCapsuleManager :
     DataflowGraphNode,
     ISideEffectApi
 {
     /// <summary>
-    /// Gets container containing capsule's data.
+    /// Initializes a new instance of the <see cref="UntypedCapsuleManager"/> class.
     /// </summary>
-    public Container Container { get; } = container;
+    /// <param name="container">Capsule container.</param>
+    protected UntypedCapsuleManager(CapsuleContainer container)
+    {
+        this.Container = container;
+    }
 
     /// <summary>
     /// Gets or sets a value indicating whether encapsulated data has been built.
     /// </summary>
-    public bool HasBuilt { get; protected set; }
+    internal bool HasBuilt { get; set; }
 
     /// <summary>
     /// Gets data of registered side effects.
     /// </summary>
-    public List<object?> SideEffectData { get; } = [];
+    internal List<object?> SideEffectData { get; } = new();
 
     /// <summary>
     /// Gets set of callbacks to dispose.
     /// </summary>
-    public HashSet<SideEffectApiCallback> ToDispose { get; } = [];
-
-    /// <inheritdoc/>
-    public override bool IsSuperPure => this.SideEffectData.Count == 0;
+    internal HashSet<SideEffectApiCallback> ToDispose { get; } = new();
 
     /// <summary>
-    /// Reads data from capsule other than the one this manages.
+    /// Gets container containing capsule's data.
     /// </summary>
-    /// <typeparam name="TRead">Type of read data.</typeparam>
-    /// <param name="otherCapsule">Other capsule whose data to read.</param>
-    /// <returns><paramref name="otherCapsule"/> data.</returns>
-    public TRead Read<TRead>(Capsule<TRead> otherCapsule)
-    {
-        var otherManager = this.Container.Manager(otherCapsule);
-        this.AddDependency(otherManager);
-        return otherManager.Data;
-    }
+    protected CapsuleContainer Container { get; }
+
+    /// <inheritdoc/>
+    protected override bool IsSuperPure => this.SideEffectData.Count == 0;
 
     /// <inheritdoc/>
     public void Rebuild() => this.BuildSelfAndDependents();
@@ -57,4 +53,17 @@ internal abstract class UntypedCapsuleManager(Container container) :
     /// <inheritdoc/>
     public void UnregisterDispose(SideEffectApiCallback callback) =>
         this.ToDispose.Remove(callback);
+
+    /// <summary>
+    /// Reads data from capsule other than the one this manages.
+    /// </summary>
+    /// <typeparam name="TRead">Type of read data.</typeparam>
+    /// <param name="otherCapsule">Other capsule whose data to read.</param>
+    /// <returns><paramref name="otherCapsule"/> data.</returns>
+    internal TRead Read<TRead>(Capsule<TRead> otherCapsule)
+    {
+        var otherManager = this.Container.ManagerOf(otherCapsule);
+        this.AddDependency(otherManager);
+        return otherManager.Data;
+    }
 }